Consolidate recipe read and write functions
