+++ websites/recipe-website/editor/src/app/(recipes)/search/page.tsx
@@ -1,4 +1,4 @@
-import getRecipes from "recipe-website-common/controller/data/readIndex";
+import { getRecipes } from "recipe-website-common/controller/data/read";
 import { RECIPES_PER_SEARCH_PAGE } from "recipe-website-common/components/SearchForm/constants";
 import { SearchForm } from "recipe-website-common/components/SearchForm";
